requirements.txt
Django==5.2.7
Pillow  # For image handling
python-dotenv==1.0.0  # For environment variable management
whitenoise==6.6.0  # For serving static files
gunicorn==21.2.0  # For production deployment
whitenoise>=6.0

# Cloudinary (media storage)
cloudinary
django-cloudinary-storage
